refaktorering. Oppdater opplysninger og infoboks for forrige behandling flyttes inn i en egen mappe og komponent: InngangsvilkårHeader
